// packages/sirius-web/frontend/sirius-web-application/src/views/edit-project/workbench-views/views-explorer/useViewsExplorerViewSubscription.types.ts
/*******************************************************************************
 * Copyright (c) 2026 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/
import { GQLTree } from '@eclipse-sirius/sirius-components-trees';

export interface GQLViewsExplorerEventData {
  viewsExplorerEvent: GQLTreeEventPayload;
}

export interface GQLTreeEventPayload {
  __typename: string;
}

export interface GQLTreeRefreshedEventPayload extends GQLTreeEventPayload {
  id: string;
  tree: GQLTree;
}

export interface UseViewsExplorerViewSubscriptionState {
  id: string;
  payload: GQLTreeEventPayload | null;
  complete: boolean;
}

export interface UseViewsExplorerViewSubscriptionValue {
  payload: GQLTreeEventPayload | null;
  loading: boolean;
  complete: boolean;
}